Remove t_Co setting
This setting was a leftover from the template I used to create the
theme. Turns out (see e.g. this stackoverflow answer -
https://stackoverflow.com/a/15378816) that setting the number of
colors displayed on the terminal from inside vim isn't a good idea, and
this should be taken care of by the terminal itself. Thus I remove
this line and apologise in case this caused any problems to anyone.