Change replace style of lightline theme
I originally had it as red bg and whitesmoke fg, as this is the way I
have it in vim-modusline, and felt it screamed "Danger!", as in I have
to be careful in that mode. However, vim-modusline also draws styles
from the used colorscheme, so it's not like my blue-mood colorscheme
had another style with a red bg. Normally it's the tomato (ff6347)
that's mostly used as the red color in this theme (the pure red being
used for selection and error messages only), so I figured that using
tomato bg and black fg would fit more with the rest of the theme,
especially as before it was the only one having white letters instead
of black.